Migrate modloader support
